// client/src/users/models/types/userType.ts
export type AddressType = {
  state?: string;
  country: string;
  city: string;
  street: string;
  houseNumber: number;
};

export type ImageType = { url?: string; alt?: string };

export type UserNameType = { first: string; middle?: string; last: string };

type UserType = {
  name: UserNameType;
  phone: string;
  email: string;
  password: string;
  isBusiness: boolean;
  image: ImageType;
  address: AddressType;
};

export type TokenType = { _id: string; isBusiness: boolean; isAdmin: boolean };

export type Login = Pick<UserType, "email" | "password">;

export type RegistrationForm = {
  first: string;
  middle: string;
  last: string;
  phone: string;
  email: string;
  password: string;
  url: string;
  alt: string;
  state: string;
  country: string;
  city: string;
  street: string;
  houseNumber: string;
  zip: string;
  isBusiness: boolean;
};

export type EditUserForm = {
  first: string;
  middle: string;
  last: string;
  phone: string;
  url: string;
  alt: string;
  state: string;
  country: string;
  city: string;
  street: string;
  houseNumber: string;
  zip: string;
};

export type RegistrationFormErrors = Partial<RegistrationForm>;
export type EditUserFormErrors = Partial<EditUserForm>

export type UserRegistered = {
  name: {
    first: string;
    middle?: string;
    last: string;
    _id?: string;
  };
  email: string;
  _id: string;
};

export type UserUpdated = {
  name: {
    first: string;
    middle?: string;
    last: string;
    _id?: string;
  };
  phone: string;
  address: AddressType;
  image: ImageType;
};

export default UserType;